customer-service/requirements.txt
Django>=4.2,<5.0
mysqlclient
djangorestframework
PyJWT
django-cors-headers
